support transfer of datasets to/from execution butler
Princeton clusters have the main repo on slow filesystems, so we transfer
files to the execution butler on a fast filesystem, and then copy them
back.

When we set up an execution butler for each job, we need to copy only the
butler configuration and registry (not the whole directory, which now
contains many many files!). And when we're done, quit Parsl activity
before starting the big merge back to the original butler.