Remove --slurmd-debug argument from Princeton site
This argument is being removed as it's causing srun to fail with the
error message:

srun: error: Use of --slurmd-debug is allowed only for root and
SlurmUser(slurm)

It is not clear when this argument became problematic, however, it must
have occurred within the last month or two. Nevertheless, use of this
argument is seemingly not required for successful pipeline execution, so
it is being removed on this ticket.