Temporary pointers to resolve GHA build issues
This temporary pointer to daf_butler on GitHub main is put in place to
resolve an emergent issue in that package causing an issue with pyarrow.
In brief, in daf_butler, PyPI depends on pyarrow, but the package
doesn't indicate that it depends on pyarrow.

This temporary edit in this package should allow for this PR to merge
now. Future edits to ctrl_bps_parsl should remove this hotfix once a
main fix has been put in place (expected 2024-01-10).

EDIT: Temporary pointers to pipe_base and ctrl_mpexec have also been
edded on this commit.

EDIT2: documenteer also temporarily pinned to v0.8.2.